Remove ability for API keys to be passed as 1st param to acct.retrieve
Old (pre-prefix) account ids and api keys are hard to differentiate. As
such old account ids would pass the `isAuthKey` tried to identify
whether a user was requesting a connected account or if they were trying
to request the account for the passed api key.

This is a "major" change in that an api key cannot be passed as a string
to the first argument. It can be passed in the options parameter as the
`api_key` value, or as a string if the user passes `null` as the account
id.

[stripe/stripe-node#194]