Remove list of available APIs from README
Removes the long list of available Node APIs from the README, and
suggests instead that people refer to our canonical API reference.

This make a somewhat convenient quick reference, but unfortunately we
don't have any plausible way of keeping them up to date so they're
especially prone to becoming inaccurate as causing confusion as seen
in #297 and #298..

Fixes #297.

Fixes #298.