Refresh README
Refreshes the README to do some general cleanup (I removed the "Version
2" update notice from 2013 and the section on the original author for
example) and to just rebuild the structure to look more like stripe-ruby
(and which will hopefully long term be a standard for all our READMEs).
For example:

* Documentation is linked first thing on top.
* Succinct installation and basic usage sections.
* All "special topics" like promises or using a Proxy get their own h3
  section under "Usage".
* "Development" section stays out of the way at the bottom.

One notable thing: I've started to eliminate the use of the word
"binding" in preference of "SDK". I spoke to @romain-stripe a bit about
this one, and we both agreed that although we're all used to "binding"
internally, more widely it's a little unusual of a word to use.