Updated utils.getDataFromArgs to support mixed options/data hash
Mixed options/data hashes may occur when a client is making a request on behalf of a Connect account(aka `stripe_account: "acct_<some account id>"`) and wishing to expand a particular field of the resource(e.g. `expand: ["charges.data.balance_transaction"]`. Meaning the hash would look like this in our figurative example:
```
{
  stripe_account: "acc_<some account id>",
  expand: ["charges.data.balance_transaction"]
}
```
Currently `utils.getDataFromArgs` does not support a mixed options/data hash. Meaning the above hash would result in the `expand` data arg being discarded instead of included. So in order to support mixed options/data hash, we instead look for keys that are not in `OPTIONS_KEYS` and pluck those from the hash. As a part of this plucking, we delete the data properties from the hash object, because they are not options that should be plucked in `utils.getOptionsFromArgs`. This way misleading warnings will not be emitted. We also change a previous `getDataFromArgs` test from warning about a mixed options/data hash to instead finding the data in a mixed options/data hash.