fix StripeConfig.apiVersion type
The type for `StripeConfig.apiVersion` seems incorrect.  It is not required, and `null` can be passed (which it even states in the comment), but as it stands right now, the only way to avoid a typescript error is to pass it _exactly_ the string `"2020-03-02"`.

This PR makes the type match what the client constructor actually accepts -- `apiVersion` is optional, and accepts a string or null. A stricter type could be created by making a union of all possible api versions, like: `'2020-03-02` | '2019-12-03'` etc, but there are so many historical api versions it seems to make more sense just to make the type `string`.